Data restructuring + rabbithole feature working
The Rabbit hole feature let's you stay on the same topic and embrace the 'rabbit hole' of whatever you stumbled upon!
For example, if you stumbled on something about "Open source", click the rabbit to go down the "Open source" rabbit hole, and all future stumbles will be about that topic. An indicator popup appears to show which rabbit hole you're on, and you can always exit by clicking on the 'x', and go back to random category mode.